GPSRealtimeHandle/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pika==1.3.1
python-dotenv==1.0.0
sqlalchemy==2.0.23
pydantic==2.5.2 